+++ packages/mobx-state-tree/__tests__/core/api.test.ts
@@ -55,6 +55,7 @@ const METHODS_AND_INTERNAL_TYPES = stringToArray(`
     getMembers,
     getPropertyMembers,
     cast,
+    castToSnapshot,
     isType,
     isArrayType,
     isFrozenType,