+++ changelog.md
@@ -1,3 +1,5 @@
+-   Fixed references to nodes being broken after the node was replaced [#1052](https://github.com/mobxjs/mobx-state-tree/issues/1052), plus speed up of reference resolving when using IDs through [#1053](https://github.com/mobxjs/mobx-state-tree/pull/1053) by [@xaviergonz](https://github.com/xaviergonz)
+
 # 3.7.0
 
 -   Middleware events now also contain `allParentIds` (chain of causing ids, from root until (excluding) current)