+++ changelog.md
@@ -1,3 +1,7 @@
+# 3.8.0
+
+-   Added a new castToSnapshot method for TypeScript and fixed some TypeScript typings not being properly detected when using SnapshotIn types.
+
 # 3.7.1
 
 -   Fixed references to nodes being broken after the node was replaced [#1052](https://github.com/mobxjs/mobx-state-tree/issues/1052), plus speed up of reference resolving when using IDs through [#1053](https://github.com/mobxjs/mobx-state-tree/pull/1053) by [@xaviergonz](https://github.com/xaviergonz)